client: refactor
