+++ client/lib/repositories/teacher.ts
@@ -1,12 +1,47 @@
 import { unstable_cache } from "next/cache"
-import { CollectionReference } from "@google-cloud/firestore";
+import {
+  CollectionReference,
+  DocumentData,
+  FirestoreDataConverter,
+  QueryDocumentSnapshot,
+  WithFieldValue,
+} from "@google-cloud/firestore";
 import { trace } from "@opentelemetry/api"
 
-import { TeacherConverter } from "@/lib/converters/teacher";
 import { Teacher } from "@/lib/models/teacher";
-import { FirestoreClient } from "@/lib/clients";
+import { FirestoreClient } from "@/lib/clients/firebase";
 
 const COLLECTION_NAME = "teachers";
+
+export const TeacherConverter: FirestoreDataConverter<Teacher> = {
+  fromFirestore(snapshot: QueryDocumentSnapshot): Teacher {
+    const data = snapshot.data();
+    return {
+      identifier: data.identifier,
+      affiliation: data.affiliation,
+      alternateName: data.alternateName,
+      description: data.description,
+      email: data.email,
+      jobTitle: data.jobTitle,
+      knowsAbout: data.knowsAbout,
+      name: data.name,
+      workLocation: data.workLocation,
+    };
+  },
+  toFirestore(modelObject: WithFieldValue<Teacher>): WithFieldValue<DocumentData> {
+    return {
+      identifier: modelObject.identifier,
+      affiliation: modelObject.affiliation,
+      alternateName: modelObject.alternateName,
+      description: modelObject.description,
+      email: modelObject.email,
+      jobTitle: modelObject.jobTitle,
+      knowsAbout: modelObject.knowsAbout,
+      name: modelObject.name,
+      workLocation: modelObject.workLocation,
+    };
+  }
+};
 
 export class TeacherRepository {
   private collection: CollectionReference<Teacher>;
